parseMail: use the list name in list-id for lists
Some lists overwrite the From name with something like "X Y via
listname" instead of using the list's name. In this case mra will set
the name to "X Y via listname" where X Y is somebody who writes the most
often to the list. This is not particulary good. If we see a "list-id"
header in the email, then any time we encounter the list's own address,
use the name in the list-id header instead of the From field.